Add a function to get disk by serial string
This function to get the disk by given serial string.
Need to check ID_SERIAL in VM if disk is attached with serial property.
